# mock_api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "GOATS Mock API",
    "description": "模拟 GOATS 对客机器人全部 20 个后端接口，供 Apifox 本地调试使用。\n\n全局 Header 参数（每个接口都需要）：\n- clientid: TL_AGENT\n- clientsecret: tltest\n- signature: clientid+timestamp+clientsalt (MD5 16位大写)\n- timestamp: 当前时间戳ms\n\n模拟错误：任意接口追加 ?_error=1 返回 400",
    "version": "1.0.0",
    "contact": {}
  },
  "servers": [
    {
      "url": "http://localhost:8099",
      "description": "Mock 本地服务"
    }
  ],
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "clientid",
        "description": "全局认证：clientid + clientsecret + signature + timestamp"
      }
    },
    "schemas": {
...
      "DifyRerankRequest": {
        "type": "object",
        "required": ["inputs", "user", "response_mode"],
        "properties": {
          "inputs": {
            "type": "object",
            "required": ["list", "keyword"],
            "properties": {
              "list": { "type": "string", "example": "[{\"windCode\":\"0200.HK\",\"insShtDesc\":\"新濠国际发展\",\"insLngDesc\":\"新濠国际发展\",\"insFamily\":\"EQUITY\",\"currency\":\"HKD\",\"exchange\":\"HKEX\"}]", "description": "JSON序列化的标的列表" },
              "keyword": { "type": "string", "example": "0200.hk", "description": "搜索关键词" }
            }
          },
          "user": { "type": "string", "example": "ai-trading-assistant" },
          "response_mode": { "type": "string", "example": "blocking", "description": "响应模式" }
        }
      },
...
    },
...
    "parameters": {
      "ErrorSimulation": {
        "name": "_error",
        "in": "query",
        "schema": { "type": "string", "enum": ["1", "true"] },
        "description": "设为1或true模拟返回400错误",
        "required": false
      }
    }
  },
  "paths": {
...
    "/v1/workflows/run": {
      "post": {
        "tags": ["DIFY"],
        "summary": "大模型rerank标的列表",
        "operationId": "difyRerank",
        "parameters": [
          { "$ref": "#/components/parameters/ErrorSimulation" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DifyRerankRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "task_id": { "type": "string", "example": "556a6022-7a4d-4e43-be8c-f12136e769b9" },
                    "workflow_run_id": { "type": "string", "example": "f44f69db-ad29-431b-a985-e8983c9e2ea8" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "string" },
                        "workflow_id": { "type": "string" },
                        "status": { "type": "string", "example": "succeeded" },
                        "outputs": {
                          "type": "object",
                          "properties": {
                            "result": { "type": "array", "items": { "type": "string" }, "example": ["0200.HK"] }
                          }
                        },
                        "error": { "type": "string", "nullable": true },
                        "elapsed_time": { "type": "number", "example": 2.783889 },
                        "total_tokens": { "type": "integer", "example": 1999 },
                        "total_steps": { "type": "integer", "example": 6 },
                        "created_at": { "type": "integer", "example": 1777535594 },
                        "finished_at": { "type": "integer", "example": 1777535597 }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}